# contracts/openapi.json
{
  "components": {
    "schemas": {
      "AdminStatsOverview": {
        "properties": {
          "practitioner_applications": {
            "$ref": "#/components/schemas/PractitionerApplicationStats"
          },
          "total_chat_messages": {
            "title": "Total Chat Messages",
            "type": "integer"
          },
          "total_chat_sessions": {
            "title": "Total Chat Sessions",
            "type": "integer"
          },
          "total_consultations": {
            "title": "Total Consultations",
            "type": "integer"
          },
          "total_guest_kundalis_all_time": {
            "title": "Total Guest Kundalis All Time",
            "type": "integer"
          },
          "total_guest_kundalis_last_7_days": {
            "title": "Total Guest Kundalis Last 7 Days",
            "type": "integer"
          },
          "total_saved_kundalis": {
            "title": "Total Saved Kundalis",
            "type": "integer"
          },
          "total_users": {
            "title": "Total Users",
            "type": "integer"
          },
          "users_last_30_days": {
            "title": "Users Last 30 Days",
            "type": "integer"
          },
          "users_last_7_days": {
            "title": "Users Last 7 Days",
            "type": "integer"
          }
        },
        "required": [
          "total_users",
          "users_last_7_days",
          "users_last_30_days",
          "total_saved_kundalis",
          "total_chat_sessions",
          "total_chat_messages",
          "total_guest_kundalis_all_time",
          "total_guest_kundalis_last_7_days",
          "practitioner_applications",
          "total_consultations"
        ],
        "title": "AdminStatsOverview",
        "type": "object"
      },
      "AdminUserItem": {
        "properties": {
          "chat_session_count": {
            "title": "Chat Session Count",
            "type": "integer"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kundali_count": {
            "title": "Kundali Count",
            "type": "integer"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "created_at",
          "kundali_count",
          "chat_session_count"
        ],
        "title": "AdminUserItem",
        "type": "object"
      },
...
      "AvakhadaOut": {
        "description": "Traditional birth attributes, all derived from the Moon.\n\n`paya` is deliberately absent: the rules in circulation disagree and give\ndifferent metals for the same chart, so the engine declines to invent one.",
        "properties": {
          "charan": {
            "description": "The Moon's nakshatra pada, 1-4",
            "title": "Charan",
            "type": "integer"
          },
          "gana": {
            "title": "Gana",
            "type": "string"
          },
          "nadi": {
            "title": "Nadi",
            "type": "string"
          },
          "nakshatra": {
            "title": "Nakshatra",
            "type": "string"
          },
          "name_syllable": {
            "description": "Syllable traditionally used to name the child",
            "title": "Name Syllable",
            "type": "string"
          },
          "sign": {
            "title": "Sign",
            "type": "string"
          },
          "sign_lord": {
            "title": "Sign Lord",
            "type": "string"
          },
          "tatva": {
            "title": "Tatva",
            "type": "string"
          },
          "varna": {
            "title": "Varna",
            "type": "string"
          },
          "vashya": {
            "title": "Vashya",
            "type": "string"
          },
          "yoni": {
            "title": "Yoni",
            "type": "string"
          },
          "yunja": {
            "title": "Yunja",
            "type": "string"
          }
        },
        "required": [
          "varna",
          "vashya",
          "yoni",
          "gana",
          "nadi",
          "sign",
          "sign_lord",
          "charan",
          "tatva",
          "nakshatra",
          "name_syllable",
          "yunja"
        ],
        "title": "AvakhadaOut",
        "type": "object"
      },
      "BirthDetailsIn": {
        "properties": {
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "latitude": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "description": "East positive",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude",
            "type": "number"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "place_label": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Place Label",
            "type": "string"
          },
          "siddhanta": {
            "default": "surya",
            "description": "Which system computes the Sun and Moon, and therefore the panchang, the dashas and the avakhada. `surya` is \u0938\u0942\u0930\u094d\u092f \u0938\u093f\u0926\u094d\u0927\u093e\u0928\u094d\u0924, what a Nepali kundali is cast from, and the default. `drik` is the modern ephemeris \u2014 better astronomy, and what AstroSage and AstroTalk publish, so use it to compare against those. The five star-planets are drik either way.",
            "enum": [
              "surya",
              "drik"
            ],
            "title": "Siddhanta",
            "type": "string"
          },
          "time": {
            "examples": [
              "08:30"
            ],
            "pattern": "^\\d{2}:\\d{2}(:\\d{2})?$",
            "title": "Time",
            "type": "string"
          },
          "time_accuracy": {
            "default": "exact",
            "enum": [
              "exact",
              "approximate",
              "unknown"
            ],
            "title": "Time Accuracy",
            "type": "string"
          },
          "tz_name": {
            "description": "IANA zone name. NOT a UTC offset: the offset for a place changes over time, and sending a fixed one silently corrupts historical charts.",
            "examples": [
              "Asia/Kathmandu"
            ],
            "title": "Tz Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "date",
          "time",
          "tz_name",
          "latitude",
          "longitude",
          "place_label"
        ],
        "title": "BirthDetailsIn",
        "type": "object"
      },
...
      "ChartOut": {
        "properties": {
          "avakhada": {
            "$ref": "#/components/schemas/AvakhadaOut"
          },
          "ayanamsa_name": {
            "title": "Ayanamsa Name",
            "type": "string"
          },
          "ayanamsa_value": {
            "title": "Ayanamsa Value",
            "type": "number"
          },
          "computed_at": {
            "format": "date-time",
            "title": "Computed At",
            "type": "string"
          },
          "dasha": {
            "$ref": "#/components/schemas/DashaOut"
          },
          "engine_version": {
            "description": "Bumped whenever a calculation rule changes. Surfacing it makes 'why did my chart change?' answerable.",
            "title": "Engine Version",
            "type": "string"
          },
          "houses": {
            "items": {
              "$ref": "#/components/schemas/HouseOut"
            },
            "title": "Houses",
            "type": "array"
          },
          "julian_day": {
            "title": "Julian Day",
            "type": "number"
          },
          "lagna_degree": {
            "title": "Lagna Degree",
            "type": "number"
          },
          "lagna_sign": {
            "title": "Lagna Sign",
            "type": "string"
          },
          "lagna_sign_index": {
            "title": "Lagna Sign Index",
            "type": "integer"
          },
          "panchang": {
            "$ref": "#/components/schemas/PanchangOut"
          },
          "planets": {
            "items": {
              "$ref": "#/components/schemas/PlanetOut"
            },
            "title": "Planets",
            "type": "array"
          },
          "siddhanta": {
            "default": "",
            "description": "Which system the positions come from. Drik (\u0926\u0943\u0915\u094d \u0938\u093f\u0926\u094d\u0927\u093e\u0928\u094d\u0924) is modern and observational; most hand-cast Nepali panchangas follow \u0938\u0942\u0930\u094d\u092f \u0938\u093f\u0926\u094d\u0927\u093e\u0928\u094d\u0924, whose Moon runs roughly a quarter degree ahead. That only changes anything near a boundary \u2014 see `panchang.near_boundary`.",
            "title": "Siddhanta",
            "type": "string"
          },
          "tribhagi": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DashaOut"
              },
              {
                "type": "null"
              }
            ],
            "description": "Vimshottari with a third removed \u2014 an 80-year cycle."
          },
          "vargas": {
            "items": {
              "$ref": "#/components/schemas/VargaChartOut"
            },
            "title": "Vargas",
            "type": "array"
          },
          "yogini": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DashaOut"
              },
              {
                "type": "null"
              }
            ],
            "description": "The eight yoginis over 36 years. `birth_lord` and each period's `lord` name a yogini, not a graha."
          }
        },
        "required": [
          "engine_version",
          "computed_at",
          "julian_day",
          "ayanamsa_name",
          "ayanamsa_value",
          "lagna_sign",
          "lagna_sign_index",
          "lagna_degree",
          "planets",
          "houses",
          "dasha",
          "panchang",
          "avakhada",
          "vargas"
        ],
        "title": "ChartOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PractitionerApplicationStats": {
        "properties": {
          "approved": {
            "title": "Approved",
            "type": "integer"
          },
          "pending": {
            "title": "Pending",
            "type": "integer"
          },
          "rejected": {
            "title": "Rejected",
            "type": "integer"
          }
        },
        "required": [
          "pending",
          "approved",
          "rejected"
        ],
        "title": "PractitionerApplicationStats",
        "type": "object"
      },
...
      "SavedKundaliIn": {
        "properties": {
          "dob": {
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
            "title": "Dob",
            "type": "string"
          },
          "gender": {
            "default": "male",
            "pattern": "^(male|female)$",
            "title": "Gender",
            "type": "string"
          },
          "lat": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "description": "East positive",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lon",
            "type": "number"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "place_name": {
            "title": "Place Name",
            "type": "string"
          },
          "tob": {
            "pattern": "^\\d{2}:\\d{2}(:\\d{2})?$",
            "title": "Tob",
            "type": "string"
          },
          "tz_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "IANA zone name. Optional only because clients that predate this field still post without it \u2014 send it. A stored UTC offset is wrong for any historical date whose zone has since changed, and the chart it produces looks entirely normal.",
            "examples": [
              "Asia/Kathmandu"
            ],
            "title": "Tz Name"
          },
          "tz_offset": {
            "deprecated": true,
            "description": "Superseded by tz_name. Retained so old clients keep working.",
            "title": "Tz Offset",
            "type": "number"
          }
        },
        "required": [
          "name",
          "dob",
          "tob",
          "lat",
          "lon",
          "tz_offset",
          "place_name"
        ],
        "title": "SavedKundaliIn",
        "type": "object"
      },
      "SavedKundaliOut": {
        "properties": {
          "birth": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BirthDetailsIn"
              },
              {
                "type": "null"
              }
            ],
            "description": "The saved birth data in the same shape `/v1/kundali` accepts, so a saved chart can be recalculated without translation. Null when the row predates `tz_name` \u2014 such a row cannot be recalculated correctly and needs a backfill rather than a guessed offset."
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "dob": {
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
            "title": "Dob",
            "type": "string"
          },
          "gender": {
            "default": "male",
            "pattern": "^(male|female)$",
            "title": "Gender",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "lat": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "description": "East positive",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lon",
            "type": "number"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "place_name": {
            "title": "Place Name",
            "type": "string"
          },
          "tob": {
            "pattern": "^\\d{2}:\\d{2}(:\\d{2})?$",
            "title": "Tob",
            "type": "string"
          },
          "tz_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "IANA zone name. Optional only because clients that predate this field still post without it \u2014 send it. A stored UTC offset is wrong for any historical date whose zone has since changed, and the chart it produces looks entirely normal.",
            "examples": [
              "Asia/Kathmandu"
            ],
            "title": "Tz Name"
          },
          "tz_offset": {
            "deprecated": true,
            "description": "Superseded by tz_name. Retained so old clients keep working.",
            "title": "Tz Offset",
            "type": "number"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "name",
          "dob",
          "tob",
          "lat",
          "lon",
          "tz_offset",
          "place_name",
          "id",
          "user_id",
          "created_at"
        ],
        "title": "SavedKundaliOut",
        "type": "object"
      },
...
      "UpdateRoleRequest": {
        "properties": {
          "role": {
            "enum": [
              "seeker",
              "practitioner",
              "admin"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role"
        ],
        "title": "UpdateRoleRequest",
        "type": "object"
      },
      "UserListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminUserItem"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "limit"
        ],
        "title": "UserListResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\nOne backend, two clients (web and Flutter).\n\n**Compatibility:** additive changes only within `/v1` \u2014 no removed or renamed\nfields, no newly-required request fields, no narrowed types. App Store users\nstay on old builds for months. Clients must treat unknown enum values as\n\"other\". See docs/architecture.md \u00a77.\n\n**Errors** are always `{\"error\": {\"code\", \"message\", \"details\"}}`. Switch on\n`code`; never parse `message`.\n",
    "title": "Nakhatra API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/stats": {
      "get": {
        "description": "Returns aggregate system statistics for the admin dashboard.\n\nNever exposes PII. Used for high-level health monitoring.",
        "operationId": "get_stats_v1_admin_stats_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminStatsOverview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get admin dashboard stats",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/users": {
      "get": {
        "operationId": "list_users_v1_admin_users_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Search by email substring",
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search by email substring",
              "title": "Search"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List users",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/users/{user_id}/role": {
      "patch": {
        "operationId": "update_user_role_v1_admin_users__user_id__role_patch",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Update User Role V1 Admin Users  User Id  Role Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update user role",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/v1/kundali": {
      "post": {
        "description": "Computes a full Vedic chart from birth details: Lahiri ayanamsa, whole-sign houses, mean nodes, Vimshottari dasha. Deterministic and synchronous \u2014 the calculation is local arithmetic against an ephemeris, not a job.\n\nSee docs/astrology-methodology.md for every methodology decision.",
        "operationId": "create_kundali_v1_kundali_post",
        "parameters": [
          {
            "description": "Dasha levels to return: 1 maha, 2 +antar, 3 +pratyantar. Defaults to 2 \u2014 the full tree is 819 periods and ~78KB, against ~12KB for two levels. Request 3 only when drilling into a specific period.",
            "in": "query",
            "name": "dasha_depth",
            "required": false,
            "schema": {
              "default": 2,
              "description": "Dasha levels to return: 1 maha, 2 +antar, 3 +pratyantar. Defaults to 2 \u2014 the full tree is 819 periods and ~78KB, against ~12KB for two levels. Request 3 only when drilling into a specific period.",
              "maximum": 3,
              "minimum": 1,
              "title": "Dasha Depth",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BirthDetailsIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChartOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate a birth chart",
        "tags": [
          "kundali"
        ]
      }
    },
...
    "/v1/vault/kundalis": {
      "get": {
        "operationId": "list_kundalis_v1_vault_kundalis_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SavedKundaliOut"
                  },
                  "title": "Response List Kundalis V1 Vault Kundalis Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Kundalis",
        "tags": [
          "vault"
        ]
      },
      "post": {
        "operationId": "save_kundali_v1_vault_kundalis_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedKundaliIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedKundaliOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Kundali",
        "tags": [
          "vault"
        ]
      }
    },
    "/v1/vault/kundalis/{kundali_id}": {
      "delete": {
        "operationId": "delete_kundali_v1_vault_kundalis__kundali_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "kundali_id",
            "required": true,
            "schema": {
              "title": "Kundali Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Delete Kundali V1 Vault Kundalis  Kundali Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Kundali",
        "tags": [
          "vault"
        ]
      }
    },
...
  }
}